refactor: add fromMaybe and separate maybe from object in util
