+++ components/util/default/.test.list
@@ -6,6 +6,7 @@
   * counter.mjs
   * either.mjs
   * function.mjs
+  * maybe.mjs
   * object.mjs
   * version.mjs
 - index.mjs
