+++ components/util/default/maybe.mjs
@@ -1,0 +1,8 @@
+export const fromMaybe = (maybe, recovery, transform) =>
+  maybe === null ? recovery : transform(maybe);
+
+export const mapMaybe = (maybe, transform) =>
+  maybe === null ? null : transform(maybe);
+
+export const mapMaybeAsync = async (maybe, transformAsync) =>
+  maybe === null ? null : await transformAsync(maybe);
